+++ backend/http/search.go
@@ -2,24 +2,31 @@ package http
 
 import (
 	"net/http"
+
 	"github.com/gtsteffaniak/filebrowser/search"
 )
 
 var searchHandler = withUser(func(w http.ResponseWriter, r *http.Request, d *data) (int, error) {
 	response := []map[string]interface{}{}
 	query := r.URL.Query().Get("query")
-	indexInfo, fileTypes := search.SearchAllIndexes(query, r.URL.Path)
-	for _,path := range(indexInfo){
-		f := fileTypes[path]
+
+	// Retrieve the User-Agent and X-Auth headers from the request
+	sessionId := r.Header.Get("SessionId")
+	indexInfo, fileTypes := search.SearchAllIndexes(query, r.URL.Path, sessionId)
+	for _, path := range indexInfo {
 		responseObj := map[string]interface{}{
-			"path"		: 	path,
+			"path": path,
+			"dir":  true,
 		}
-		for filterType,_ := range(f) {
-			if f[filterType] {
-				responseObj[filterType] = f[filterType]
+		if _, ok := fileTypes[path]; ok {
+			responseObj["dir"] = false
+			for filterType, value := range fileTypes[path] {
+				if value {
+					responseObj[filterType] = value
+				}
 			}
 		}
-		response = append(response,responseObj)
+		response = append(response, responseObj)
 	}
 	return renderJSON(w, r, response)
 })