+++ frontend/src/components/Search.vue
@@ -5,8 +5,8 @@
         <i class="material-icons">close</i>
       </button>
       <i v-else class="material-icons">search</i>
-      <input type="text" @keyup.exact="keyup" @input="submit" ref="input" :autofocus="active" v-model.trim="value"
-        :aria-label="$t('search.search')" :placeholder="$t('search.search')" />
+      <input class="main-input" type="text" @keyup.exact="keyup" @input="submit" ref="input" :autofocus="active"
+        v-model.trim="value" :aria-label="$t('search.search')" :placeholder="$t('search.search')" />
     </div>
     <div v-if="isMobile && active" id="result" :class="{ hidden: !active }" ref="result">
       <div id="result-list">
@@ -31,14 +31,19 @@
         <p v-show="isEmpty && isRunning" id="renew">
           <i class="material-icons spin">autorenew</i>
         </p>
-        <p v-show="isEmpty && !isRunning">{{ text }}</p>
+        <div v-show="isEmpty && !isRunning">
+          <div class="searchPrompt" v-show="isEmpty && !isRunning">
+            <p>{{ noneMessage }}</p>
+          </div>
+        </div>
         <template v-if="isEmpty">
-          <template v-if="value.length === 0">
+          <button class="mobile-boxes" v-if="value.length === 0 && !showBoxes " @click="resetSearchFilters()" >Reset filters</button>
+          <template v-if="value.length === 0 && showBoxes ">
             <div class="boxes">
               <h3>{{ $t("search.types") }}</h3>
               <div>
-                <div tabindex="0" v-for="(v, k) in boxes" :key="k" role="button" @click="init('type:' + k)"
-                  :aria-label="v.label">
+                <div class="mobile-boxes" tabindex="0" v-for="(v, k) in boxes" :key="k" role="button"
+                  @click="addToTypes('type:' + k)" :aria-label="v.label">
                   <i class="material-icons">{{ v.icon }}</i>
                   <p>{{ v.label }}</p>
                 </div>
@@ -50,9 +55,47 @@
     </div>
     <div v-if="!isMobile && active" id="result-desktop" ref="result">
       <div id="result-list">
-        <div class="button" style="width: 100%">
+        <div class="button fluid">
           Search Context: {{ getContext(this.$route.path) }}
         </div>
+        <template>
+          <p v-show="isEmpty && isRunning" id="renew">
+            <i class="material-icons spin">autorenew</i>
+          </p>
+          <div class="searchPrompt" v-show="isEmpty && !isRunning">
+            <p>{{ noneMessage }}</p>
+            <div class="helpButton" @click="toggleHelp()">Help</div>
+          </div>
+          <div class="helpText" v-if="showHelp">
+            <p>Search occurs on each character you type (3 character minimum for search terms).</p>
+            <p><b>The index:</b> Search utilizes the index which automatically gets updated on the configured interval
+              (default: 5 minutes).
+              Searching when the program has just started may result in incomplete results.</p>
+            <p><b>Filter by type:</b> You can have multiple type filters by adding <code>type:condition</code> followed by
+              search terms.</p>
+            <p><b>Multiple Search terms:</b> Additional terms separated by <code>|</code>,
+              for example <code>"test|not"</code> searches for both terms independently.</p>
+            <p><b>File size:</b> Searching files by size may have significantly longer search times.</p>
+          </div>
+          <template>
+            <ButtonGroup :buttons="folderSelect" @button-clicked="addToTypes" @remove-button-clicked="removeFromTypes"
+              @disableAll="folderSelectClicked()" @enableAll="resetButtonGroups()" />
+            <ButtonGroup :buttons="typeSelect" @button-clicked="addToTypes" @remove-button-clicked="removeFromTypes"
+              :isDisabled="isTypeSelectDisabled" />
+            <div class="sizeConstraints">
+              <div class="sizeInputWrapper">
+                <p>Smaller Than:</p>
+                <input class="sizeInput" v-model="smallerThan" type="text" placeholder="number">
+                <p>MB</p>
+              </div>
+              <div class="sizeInputWrapper">
+                <p>Larger Than:</p>
+                <input class="sizeInput" v-model="largerThan" type="text" placeholder="number">
+                <p>MB</p>
+              </div>
+            </div>
+          </template>
+        </template>
         <ul v-show="results.length > 0">
           <li v-for="(s, k) in results" :key="k" @click.stop.prevent="navigateTo(s.url)" style="cursor: pointer">
             <router-link to="#" event="">
@@ -68,23 +111,89 @@
             </router-link>
           </li>
         </ul>
-        <template>
-          <p v-show="isEmpty && isRunning" id="renew">
-            <i class="material-icons spin">autorenew</i>
-          </p>
-          <p v-show="isEmpty && !isRunning">{{ text }}</p>
-          <template>
-            <div v-show="results.length == 0" class="boxes">
-              <ButtonGroup :buttons="folderSelect" @button-clicked="init" @remove-button-clicked="removeInit" />
-              <ButtonGroup :buttons="typeSelect" @button-clicked="init" @remove-button-clicked="removeInit" />
-              <ButtonGroup :buttons="sizeSelect" @button-clicked="init" @remove-button-clicked="removeInit" />
-            </div>
-          </template>
-        </template>
       </div>
     </div>
   </div>
 </template>
+
+<style>
+.main-input {
+  width: 100%
+}
+
+.mobile-boxes {
+  cursor: pointer;
+  overflow: hidden;
+  margin-bottom: 1em;
+  background: var(--blue);
+  color: white;
+  padding: 1em;
+  border-radius: 1em;
+  text-align: center;
+}
+/* Hiding scrollbar for Chrome, Safari and Opera */
+.mobile-boxes::-webkit-scrollbar {
+    display: none;
+}
+
+/* Hiding scrollbar for IE, Edge and Firefox */
+.mobile-boxes {
+  scrollbar-width: none;  /* Firefox */
+  -ms-overflow-style: none;  /* IE and Edge */
+}
+.helpText {
+  padding: 1em
+}
+
+.sizeConstraints {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  align-content: center;
+  margin: 1em;
+  justify-content: center;
+}
+
+.sizeInput {
+  height: 100%;
+  text-align: center;
+  width: 5em;
+  border-radius: 1em;
+  padding: 1em;
+
+  border: solid !important;
+}
+
+.sizeInputWrapper {
+  border-radius: 1em;
+  margin-left: 0.5em;
+  margin-right: 0.5em;
+  border-style: groove;
+  display: flex;
+  background-color: rgb(245, 245, 245);
+  padding: .25em;
+  height: 3em;
+  align-items: center;
+}
+
+.helpButton {
+  position: absolute;
+  right: 10px;
+  cursor: pointer;
+  text-align: center;
+  background: rgb(211, 211, 211);
+  padding: 0.25em;
+  border-radius: 0.25em;
+}
+
+.searchPrompt {
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+}
+</style>
 
 <script>
 import ButtonGroup from "./ButtonGroup.vue";
@@ -99,8 +208,6 @@ var boxes = {
   audio: { label: "audio files", icon: "volume_up" },
   video: { label: "videos", icon: "movie" },
   doc: { label: "documents", icon: "picture_as_pdf" },
-  "larger=100": { label: "larger than 100MB", icon: "arrow_forward_ios" },
-  "smaller=100": { label: "smaller than 100MB ", icon: "arrow_back_ios" },
 };
 
 export default {
@@ -110,6 +217,12 @@ export default {
   name: "search",
   data: function () {
     return {
+      largerThan: "",
+      smallerThan: "",
+      noneMessage: "Start typing 3 or more characters to begin searching.",
+      searchTypes: "",
+      isTypeSelectDisabled: false,
+      showHelp: false,
       folderSelect: [
         { label: "Only Folders", value: "type:folder" },
         { label: "Only Files", value: "type:file" },
@@ -120,10 +233,6 @@ export default {
         { label: "Videos", value: "type:video" },
         { label: "Documents", value: "type:docs" },
         { label: "Archives", value: "type:archive" },
-      ],
-      sizeSelect: [
-        { label: "Smaller than 100MB", value: "type:smaller=100" },
-        { label: "Larger than 100MB", value: "type:larger=100" },
       ],
       value: "",
       width: window.innerWidth,
@@ -143,7 +252,8 @@ export default {
         }
 
         document.body.style.overflow = "auto";
-        this.reset();
+        this.ongoing = false;
+        this.results = [];
         this.value = "";
         this.active = false;
         this.$refs.input.blur();
@@ -155,13 +265,17 @@ export default {
     },
     value() {
       if (this.results.length) {
-        this.reset();
+        this.ongoing = false;
+        this.results = [];
       }
     },
   },
   computed: {
     ...mapState(["user", "show"]),
     ...mapGetters(["isListing"]),
+    showBoxes() {
+      return this.searchTypes == "";
+    },
     boxes() {
       return boxes;
     },
@@ -182,7 +296,10 @@ export default {
     },
     isRunning() {
       return this.ongoing;
-    }
+    },
+    searchHelp() {
+      return this.showHelp
+    },
   },
   mounted() {
     window.addEventListener("resize", this.handleResize);
@@ -206,21 +323,22 @@ export default {
       if (!str.includes("/")) {
         return "";
       }
-      let parts = str.replace(/\/$/, "").split("/");
+      let parts = str.replace(/(\/$|^\/)/, "").split("/"); //remove first and last slash
       parts.pop();
-      return parts.join("/") + "/";
+      parts = parts.join("/") + "/"
+      if (str.endsWith("/")) {
+        parts = "/" + parts // weird rtl bug
+      }
+      return parts;
     },
     baseName(str) {
-      let parts = str.split("/");
-      if (str.endsWith("/")) {
-        return parts[parts.length - 2] + "/";
-      } else {
-        return parts[parts.length - 1];
-      }
+      let parts = str.replace(/(\/$|^\/)/, "").split("/")
+      return parts.pop()
     },
     ...mapMutations(["showHover", "closeHovers", "setReload"]),
     open() {
       this.showHover("search");
+      this.showBoxes = true;
     },
     close(event) {
       event.stopPropagation();
@@ -233,44 +351,64 @@ export default {
       }
       this.results.length === 0;
     },
-    init(string) {
+    addToTypes(string) {
+      if (this.searchTypes.includes(string)) {
+        return true
+      }
       if (string == null || string == "") {
         return false
       }
-      this.value = `${string} ${this.value}`;
+      this.searchTypes = this.searchTypes + string + " "
+    },
+    resetSearchFilters(){
+      this.searchTypes= "";
+    },
+    removeFromTypes(string) {
+      if (string == null || string == "") {
+        return false
+      }
+      this.searchTypes = this.searchTypes.replace(string + " ", "");
       if (this.isMobile) {
         this.$refs.input.focus();
       }
     },
-    removeInit(string) {
-      if (string == null || string == "") {
-        return false
-      }
-      this.value = this.value.replace(string + " ", "");
-      if (this.isMobile) {
-        this.$refs.input.focus();
-      }
+    folderSelectClicked() {
+      this.isTypeSelectDisabled = true;  // Disable the other ButtonGroup
     },
-    reset() {
-      this.ongoing = false;
-      this.resultsCount = 50;
-      this.results = [];
+    resetButtonGroups() {
+      this.isTypeSelectDisabled = false;
     },
     async submit(event) {
+      this.showHelp = false;
       event.preventDefault();
-      const words = this.value.split(" ").filter((word) => word.length < 3);
-      if (this.value === "" || words.length > 0) {
-        return;
+      if (this.value === "" || this.value.length < 3) {
+        this.ongoing = false;
+        this.results = [];
+        this.noneMessage = "Not enough characters to search (min 3)"
+        return
+      }
+      let searchTypesFull = this.searchTypes
+      if (this.largerThan != "") {
+        searchTypesFull = searchTypesFull + "type:largerThan=" + this.largerThan + " "
+      }
+      if (this.smallerThan != "") {
+        searchTypesFull = searchTypesFull + "type:smallerThan=" + this.smallerThan + " "
       }
       let path = this.$route.path;
       this.ongoing = true;
       try {
-        this.results = await search(path, this.value);
+        this.results = await search(path, searchTypesFull + this.value);
       } catch (error) {
         this.$showError(error);
       }
+      if (this.results.length == 0 && this.ongoing == false) {
+        this.noneMessage = "No results found in indexed search."
+      }
       this.ongoing = false;
     },
+    toggleHelp() {
+      this.showHelp = !this.showHelp
+    }
   },
 };
 </script>