+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+# v0.1.4
+ - various UI fixes
+   - Added download button back to toolbar
+   - Added upload button to side menu
+   - breadcrumb spacing fix
+   - Added "compact" view option
+   - fixed slash issue with css rtl logic
+ - various backend fixes
+   - search has a sessionId attached so searches don't collide
+   - search no longer searches by word with spaces, includes space in searches
+   - prepared for full json configuration
+ - made size search work for smaller and larger
+ - made search types not show up in search bar when used
+
 ## v0.1.3
 
  - improved styling, colors, transparency, blur
@@ -11,7 +25,7 @@ All notable changes to this project will be documented in this file. See [standa
    - search
    - toggle view
  - Changed desktop search style and included additional search options.
-   
+
 ## v0.1.2
 
  - Updated UI to use search features better
@@ -19,7 +33,7 @@ All notable changes to this project will be documented in this file. See [standa
    - Better icons with colors
    - GUI styling
  - Improved search performance
-   
+
 ## v0.1.1
 
  - Improved search with indexing